# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Microblogging API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/user/{userId}:
    get:
      operationId: getPostsByUser
      tags: [posts]
      summary: Get all posts by a specific user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: User's posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile by ID or username
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
        "400":
          description: Invalid request (e.g. username taken)
...
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List standalone pages
      parameters:
        - in: query
          name: includeDrafts
          schema:
            type: string
          description: When `1`, also returns drafts (owner only).
      responses:
        "200":
          description: Pages list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagesList"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new page (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageBody"
      responses:
        "201":
          description: Created page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "400":
          description: Invalid body / reserved slug / slug collision
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /pages/{slug}:
    get:
      operationId: getPageBySlug
      tags: [pages]
      summary: Fetch a single page by slug (drafts 404 for non-owners)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "404":
          description: Not found
  /pages/{id}:
    patch:
      operationId: updatePage
      tags: [pages]
      summary: Update a page (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageBody"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "400":
          description: Invalid body / reserved slug / slug collision
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page (owner only). Cascade-deletes its nav row.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    Category:
      type: object
      description: Owner-managed taxonomy entry. Slug is the canonical identifier in URLs.
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - createdAt
        - updatedAt
...
    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        contentFormat:
          type: string
          enum: [plain, html]
        commentCount:
          type: integer
        sourceFeedId:
          type: integer
          nullable: true
          description: ID of the feed_sources row that imported this post (PESOS); null for owner-authored posts.
        sourceFeedName:
          type: string
          nullable: true
          description: Display name of the originating feed source (joined from feed_sources). Null for owner-authored posts.
        sourceCanonicalUrl:
          type: string
          nullable: true
          description: Permalink of the post on its origin site (PESOS attribution).
        categories:
          type: array
          description: Categories this post belongs to (owner-managed taxonomy).
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - authorName
        - content
        - contentFormat
        - commentCount
        - categories
        - createdAt
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        socialLinks:
          type: object
          additionalProperties:
            type: string
          nullable: true
        postCount:
          type: integer
        theme:
          type: string
          enum:
            - bauhaus
            - traditional
            - minimalist
            - academic
            - airy
            - nature
            - comfort
            - audacious
            - artistic
          nullable: true
        palette:
          type: string
          enum:
            - bauhaus
            - monochrome
            - newsprint
            - ocean
            - forest
            - sunset
            - sepia
            - high-contrast
            - pastel
          nullable: true
        colorBackground:
          type: string
          nullable: true
        colorForeground:
          type: string
          nullable: true
        colorBackgroundDark:
          type: string
          nullable: true
        colorForegroundDark:
          type: string
          nullable: true
        colorPrimary:
          type: string
          nullable: true
        colorPrimaryForeground:
          type: string
          nullable: true
        colorSecondary:
          type: string
          nullable: true
        colorSecondaryForeground:
          type: string
          nullable: true
        colorAccent:
          type: string
          nullable: true
        colorAccentForeground:
          type: string
          nullable: true
        colorMuted:
          type: string
          nullable: true
        colorMutedForeground:
          type: string
          nullable: true
        colorDestructive:
          type: string
          nullable: true
        colorDestructiveForeground:
          type: string
          nullable: true
      required:
        - id
        - name
        - postCount
    UpdateUserProfileBody:
      type: object
      description: |
        Partial update to the current user's profile. Theme columns
        (`theme`, `palette`, and the 14 `color*` fields) accept an explicit
        `null` to clear the user's customization for that column, which
        makes their profile fall back to the site-wide theme value.
      properties:
        username:
          type: string
          pattern: "^[a-zA-Z0-9_]{3,30}$"
        bio:
          type: string
          maxLength: 500
        website:
          type: string
          format: uri
        socialLinks:
          type: object
          additionalProperties:
            type: string
        theme:
          type: string
          nullable: true
          enum:
            - bauhaus
            - traditional
            - minimalist
            - academic
            - airy
            - nature
            - comfort
            - audacious
            - artistic
        palette:
          type: string
          nullable: true
          enum:
            - bauhaus
            - monochrome
            - newsprint
            - ocean
            - forest
            - sunset
            - sepia
            - high-contrast
            - pastel
        colorBackground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorForeground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorBackgroundDark:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorForegroundDark:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorPrimary:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorPrimaryForeground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorSecondary:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorSecondaryForeground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorAccent:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorAccentForeground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorMuted:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorMutedForeground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorDestructive:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
        colorDestructiveForeground:
          type: string
          nullable: true
          maxLength: 32
          pattern: "^[0-9]{1,3}(\\.[0-9]+)? [0-9]{1,3}(\\.[0-9]+)?% [0-9]{1,3}(\\.[0-9]+)?%$"
...
    Page:
      type: object
      description: |
        A standalone CMS page rendered at `/p/:slug`. Title and slug are
        independent fields; the slug is the URL key. `contentFormat` is
        always `'html'` today (the column exists for future forks that
        want plain pages). Drafts return 404 to non-owners.
      properties:
        id:
          type: integer
        slug:
          type: string
          maxLength: 96
        title:
          type: string
          maxLength: 255
        content:
          type: string
        contentFormat:
          type: string
          enum: [html]
        status:
          type: string
          enum: [draft, published]
        showInNav:
          type: boolean
        authorUserId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - content
        - contentFormat
        - status
        - showInNav
        - createdAt
        - updatedAt
    PagesList:
      type: object
      properties:
        pages:
          type: array
          items:
            $ref: "#/components/schemas/Page"
      required:
        - pages
    CreatePageBody:
      type: object
      properties:
        slug:
          type: string
          maxLength: 96
        title:
          type: string
          maxLength: 255
        content:
          type: string
        status:
          type: string
          enum: [draft, published]
          default: draft
        showInNav:
          type: boolean
          default: true
      required:
        - slug
        - title
        - content
    UpdatePageBody:
      type: object
      properties:
        slug:
          type: string
          maxLength: 96
        title:
          type: string
          maxLength: 255
        content:
          type: string
        status:
          type: string
          enum: [draft, published]
        showInNav:
          type: boolean
...